+++ R/Connection.R
@@ -20,6 +20,7 @@ class_cache <- new.env(parent = emptyenv())
 AthenaConnection <- function(aws_access_key_id = NULL,
                              aws_secret_access_key = NULL,
                              aws_session_token = NULL,
+                             catalog_name = NULL,
                              schema_name = NULL,
                              work_group = NULL,
                              poll_interval = NULL,
@@ -74,11 +75,16 @@ AthenaConnection <- function(aws_access_key_id = NULL,
   # return a subset of api function to reduce object size
   ptr_ll <- list(
     Athena = Athena[c(
-      ".internal", "start_query_execution", "get_query_execution", "stop_query_execution", "get_query_results",
-      "get_work_group", "list_work_groups", "update_work_group", "create_work_group", "delete_work_group"
+      ".internal", "start_query_execution", "get_query_execution",
+      "stop_query_execution", "get_query_results", "get_work_group",
+      "list_work_groups", "list_data_catalogs", "list_databases",
+      "get_table_metadata", "update_work_group", "create_work_group",
+      "delete_work_group"
     )],
-    S3 = S3[c(".internal", "put_object", "get_object", "delete_object", "delete_objects", "list_objects_v2")],
-    glue = glue[c(".internal", "get_databases", "get_tables", "get_table", "delete_table")]
+    S3 = S3[c(
+      ".internal", "put_object", "get_object", "download_file",
+      "delete_object", "delete_objects", "list_objects_v2"
+    )]
   )
   s3_staging_dir <- s3_staging_dir %||% get_aws_env("AWS_ATHENA_S3_STAGING_DIR")
 
@@ -91,7 +97,8 @@ AthenaConnection <- function(aws_access_key_id = NULL,
   }
   info <- list(
     profile_name = prof_name, s3_staging = s3_staging_dir,
-    dbms.name = schema_name, work_group = work_group %||% "primary",
+    db.catalog = catalog_name, dbms.name = schema_name,
+    work_group = work_group %||% "primary",
     poll_interval = poll_interval, encryption_option = encryption_option,
     kms_key = kms_key, expiration = aws_expiration,
     timezone = character(),
@@ -445,6 +452,7 @@ setMethod(
 #' Returns the unquoted names of Athena tables accessible through this connection.
 #' @name dbListTables
 #' @inheritParams DBI::dbListTables
+#' @param catalog Athena catalog, default set to NULL to return all tables from all Athena catalogs
 #' @param schema Athena schema, default set to NULL to return all tables from all Athena schemas.
 #'               Note: The use of DATABASE and SCHEMA is interchangeable within Athena.
 #' @aliases dbListTables
@@ -473,24 +481,31 @@ NULL
 #' @export
 setMethod(
   "dbListTables", "AthenaConnection",
-  function(conn, schema = NULL, ...) {
+  function(conn, catalog = NULL, schema = NULL, ...) {
     con_error_msg(conn, msg = "Connection already closed.")
-    glue <- conn@ptr$glue
-    if (is.null(schema)) {
-      schema <- get_databases(glue)
-    }
-    tryCatch(
-      {
-        output <- lapply(schema, function(i) get_table_list(glue = glue, schema = i))
-      },
-      error = function(cond) NULL
-    )
-    return(
-      vapply(
-        unlist(output, recursive = FALSE),
-        function(y) y$Name,
-        FUN.VALUE = character(1)
+    cat_filter <- ""
+    db_filter <- ""
+    if (!is.null(catalog)) {
+      cat_filter <- sprintf(
+        "and lower(table_catalog) in ('%s')",
+        paste(tolower(catalog), collapse = "', '")
       )
+    }
+    if (!is.null(schema)) {
+      db_filter <- sprintf(
+        "and lower(table_schema) in ('%s')",
+        paste(tolower(schema), collapse = "', '")
+      )
+    }
+    
+    query <- "select
+      table_name
+    from information_schema.tables
+    where table_schema != 'information_schema' %s %s
+    order by table_catalog, table_schema
+    "
+    output <- suppressMessages(
+      dbGetQuery(conn, sprintf(query, cat_filter, db_filter))[["table_name"]]
     )
   }
 )
@@ -500,6 +515,7 @@ setMethod(
 #' Method to get Athena schema, tables and table types return as a data.frame
 #' @name dbGetTables
 #' @inheritParams DBI::dbListTables
+#' @param catalog Athena catalog, default set to NULL to return all tables from all Athena catalogs
 #' @param schema Athena schema, default set to NULL to return all tables from all Athena schemas.
 #'               Note: The use of DATABASE and SCHEMA is interchangeable within Athena.
 #' @aliases dbGetTables
@@ -532,20 +548,36 @@ setGeneric("dbGetTables", function(conn, ...) standardGeneric("dbGetTables"))
 #' @export
 setMethod(
   "dbGetTables", "AthenaConnection",
-  function(conn, schema = NULL, ...) {
+  function(conn, catalog = NULL, schema = NULL, ...) {
     con_error_msg(conn, msg = "Connection already closed.")
-    glue <- conn@ptr$glue
-    if (is.null(schema)) {
-      schema <- get_databases(glue)
+    
+    cat_filter <- ""
+    db_filter <- ""
+    if (!is.null(catalog)) {
+      cat_filter <- sprintf(
+        "and lower(table_catalog) in ('%s')",
+        paste(tolower(catalog), collapse = "', '")
+      )
     }
-    tryCatch(
-      {
-        output <- lapply(schema, function(i) get_table_list(glue = glue, schema = i))
-      },
-      error = function(cond) NULL
+    if (!is.null(schema)) {
+      db_filter <- sprintf(
+        "and lower(table_schema) in ('%s')",
+        paste(tolower(schema), collapse = "', '")
+      )
+    }
+    
+    query <- "select
+      table_catalog as Catalog,
+      table_schema as Schema,
+      table_name as TableName,
+      table_type as TableType
+    from information_schema.tables
+    where table_schema != 'information_schema' %s %s
+    order by table_catalog, table_schema
+    "
+    output <- suppressMessages(
+      dbGetQuery(conn, sprintf(query, cat_filter, db_filter))
     )
-    output <- rbindlist(unlist(output, recursive = FALSE), use.names = TRUE)
-    setnames(output, new = c("Schema", "TableName", "TableType"))
     return(output)
   }
 )
@@ -591,13 +623,14 @@ setMethod(
     con_error_msg(conn, msg = "Connection already closed.")
     ll <- db_detect(conn, name)
     retry_api_call(
-      output <- conn@ptr$glue$get_table(
+      output <- conn@ptr$Athena$get_table_metadata(
+        CatalogName = ll[["db.catalog"]],
         DatabaseName = ll[["dbms.name"]],
-        Name = ll[["table"]]
-      )$Table
+        TableName = ll[["table"]]
+      )$TableMetadata
     )
 
-    col_names <- vapply(output$StorageDescriptor$Columns, function(y) y$Name, FUN.VALUE = character(1))
+    col_names <- vapply(output$Columns, function(y) y$Name, FUN.VALUE = character(1))
     partitions <- vapply(output$PartitionKeys, function(y) y$Name, FUN.VALUE = character(1))
     c(col_names, partitions)
   }
@@ -643,30 +676,43 @@ setMethod(
   function(conn, name, ...) {
     con_error_msg(conn, msg = "Connection already closed.")
     ll <- db_detect(conn, name)
-
-    for (i in seq_len(athena_option_env$retry)) {
-      resp <- tryCatch(conn@ptr$glue$get_table(DatabaseName = ll[["dbms.name"]], Name = ll[["table"]]),
-        error = function(e) e
-      )
-
-      # exponential step back if error and not expected error
-      if (inherits(resp, "error") && !grepl(".*table.*not.*found.*", resp, ignore.case = T)) {
-        backoff_len <- runif(n = 1, min = 0, max = (2^i - 1))
-
-        info_msg(resp, "Request failed. Retrying in ", round(backoff_len, 1), " seconds...")
-
-        Sys.sleep(backoff_len)
-      } else {
-        break
-      }
+    for (i in seq_len(athena_option_env$retry + 1)) {
+      resp <- tryCatch({
+        conn@ptr$Athena$get_table_metadata(
+          CatalogName = ll[["db.catalog"]],
+          DatabaseName = ll[["dbms.name"]],
+          TableName = ll[["table"]]
+        )
+      }, error = function(err) {
+        err_msg = err$message
+        if(i == (athena_option_env$retry + 1)) {
+          stop(err_msg, call. = F)
+        } 
+        if (!grepl("EntityNotFoundException|Cannot.*find.*catalog", err_msg)) {
+          backoff <- 2**i * 0.5
+          info_msg(err_msg)
+          info_msg(
+            paste("Request failed. Retrying in", backoff, "seconds...")
+          )
+          Sys.sleep(backoff)
+          return(err)
+        }
+        return(FALSE)
+      })
+      if (inherits(resp, "error")) next
+      break
     }
-
-
-    if (inherits(resp, "error") && !grepl(".*table.*not.*found.*", resp, ignore.case = T)) stop(resp)
-
-    !grepl(".*table.*not.*found.*", resp[1], ignore.case = T)
+    return(is.list(resp))
   }
 )
+
+#' @rdname dbExistsTable
+#' @export
+setMethod(
+  "dbExistsTable", c("AthenaConnection", "Id"),
+  function(conn, name, ...) {
+    dbExistsTable(conn, dbQuoteIdentifier(conn, name), ...)
+})
 
 #' Remove table from Athena
 #'
@@ -678,7 +724,8 @@ setMethod(
 #' @param confirm Allows for S3 files to be deleted without the prompt check. It is recommend to leave this set to \code{FALSE}
 #'                   to avoid deleting other S3 files when the table's definition points to the root of S3 bucket.
 #' @seealso \code{\link[DBI]{dbRemoveTable}}
-#' @note If you are having difficulty removing AWS S3 files please check if the AWS S3 location following AWS best practises: \href{https://docs.aws.amazon.com/athena/latest/ug/tables-location-format.html}{Table Location in Amazon S3}
+#' @note If you are having difficulty removing AWS S3 files please check if the
+#' AWS S3 location following AWS best practises: \href{https://docs.aws.amazon.com/athena/latest/ug/tables-location-format.html}{Table Location in Amazon S3}
 #' @examples
 #' \dontrun{
 #' # Note:
@@ -717,17 +764,14 @@ setMethod(
     )
     ll <- db_detect(conn, name)
 
-    TableType <- conn@ptr$glue$get_table(DatabaseName = ll[["dbms.name"]], Name = ll[["table"]])[["Table"]][["TableType"]]
-
-    if (delete_data && TableType == "EXTERNAL_TABLE") {
-      tryCatch(
-        s3_path <- split_s3_uri(
-          conn@ptr$glue$get_table(
-            DatabaseName = ll[["dbms.name"]],
-            Name = ll[["table"]]
-          )[["Table"]][["StorageDescriptor"]][["Location"]]
-        )
-      )
+    TableMeta <- conn@ptr$Athena$get_table_metadata(
+      CatalogName = ll[["db.catalog"]],
+      DatabaseName = ll[["dbms.name"]],
+      TableName = ll[["table"]]
+    )[["TableMetadata"]]
+    
+    if (delete_data && TableMeta$TableType == "EXTERNAL_TABLE") {
+      s3_path <- split_s3_uri(TableMeta$Parameters$location)
       # Detect if key ends with "/" or if it has ".": https://github.com/DyfanJones/noctua/issues/125
       if (!grepl("\\.|/$", s3_path$key)) {
         s3_path[["key"]] <- sprintf("%s/", s3_path[["key"]])
@@ -770,15 +814,23 @@ setMethod(
         ), call. = F)
       }
     }
-
-    # use glue to remove table from glue catalog
-    conn@ptr$glue$delete_table(DatabaseName = ll[["dbms.name"]], Name = ll[["table"]])
+    
+    # Drop Athena table
+    dbExecute(conn, sprintf('DROP TABLE IF EXISTS `%s`', paste(ll, collapse = '`.`')))
 
     if (!delete_data) info_msg("Only Athena table has been removed.")
     on_connection_updated(conn, ll[["table"]])
     invisible(TRUE)
   }
 )
+
+#' @rdname dbRemoveTable
+#' @export
+setMethod(
+  "dbRemoveTable", c("AthenaConnection", "Id"),
+  function(conn, name, delete_data = TRUE, confirm = FALSE, ...) {
+    dbRemoveTable(conn, dbQuoteIdentifier(conn, name), ...)
+})
 
 #' Send query, retrieve results and then clear result set
 #'
@@ -947,7 +999,12 @@ setMethod(
     con_error_msg(conn, msg = "Connection already closed.")
     stopifnot(is.logical(.format))
     ll <- db_detect(conn, name)
-    dt <- dbGetQuery(conn, paste0("SHOW PARTITIONS ", ll[["dbms.name"]], ".", ll[["table"]]))
+    dt <- dbGetQuery(
+      conn,
+      sprintf(
+        "SHOW PARTITIONS %s", paste(ll, collapse = ".")
+      )
+    )
 
     if (.format) {
       # ensure returning format is data.table
@@ -1015,7 +1072,7 @@ setMethod(
   function(conn, name, ...) {
     con_error_msg(conn, msg = "Connection already closed.")
     ll <- db_detect(conn, name)
-    SQL(paste0(dbGetQuery(conn, paste0("SHOW CREATE TABLE ", ll[["dbms.name"]], ".", ll[["table"]]), unload = FALSE)[[1]], collapse = "\n"))
+    SQL(paste0(dbGetQuery(conn, sprintf("SHOW CREATE TABLE `%s`", paste0(ll, collapse = "`.`")), unload = FALSE)[[1]], collapse = "\n"))
   }
 )
 
@@ -1111,16 +1168,52 @@ setMethod(
     ins <- if (inherits(obj, "SQL")) {
       obj
     } else {
-      paste0("SELECT * FROM ", paste0('"', unlist(strsplit(obj, "\\.")), '"', collapse = "."))
+      sprintf('SELECT * FROM "%s"', paste(db_detect(conn, obj), collapse='"."'))
     }
 
     tt_sql <- paste0(
-      "CREATE TABLE ", paste0('"', unlist(strsplit(name, "\\.")), '"', collapse = "."),
-      " ", ctas_sql_with(partition, s3.location, file.type, compress), "AS ",
+      'CREATE TABLE "', paste(db_detect(conn, name), collapse='"."'),
+      '" ', ctas_sql_with(partition, s3.location, file.type, compress), "AS ",
       ins, "\nWITH", with_data, "DATA", ";"
     )
     res <- dbExecute(conn, tt_sql, unload = FALSE)
-    dbClearResult(res)
+    on.exit(dbClearResult(res))
     return(invisible(TRUE))
   }
 )
+
+# Unable to get Athena equivalents, for the time being will pass these methods.
+
+#' @rdname AthenaConnection
+#' @inheritParams DBI::dbBegin
+#' @export
+setMethod(
+  "dbBegin", "AthenaConnection",
+  function(conn, ...) {
+    con_error_msg(conn, msg = "Connection already closed.")
+    invisible(TRUE)
+})
+
+# https://github.com/laughingman7743/PyAthena/blob/f4b21a0b0f501f5c3504698e25081f491a541d4e/pyathena/connection.py#L281C1-L282
+
+#' @rdname AthenaConnection
+#' @inheritParams DBI::dbCommit
+#' @export
+setMethod(
+  "dbCommit", "AthenaConnection",
+  function(conn, ...) {
+    con_error_msg(conn, msg = "Connection already closed.")
+    invisible(TRUE)
+  })
+
+
+# https://github.com/laughingman7743/PyAthena/blob/f4b21a0b0f501f5c3504698e25081f491a541d4e/pyathena/connection.py#L284-L285
+#' @rdname AthenaConnection
+#' @inheritParams DBI::dbRollback
+#' @export
+setMethod(
+  "dbRollback", "AthenaConnection",
+  function(conn, ...) {
+    con_error_msg(conn, msg = "Connection already closed.")
+    invisible(TRUE)
+})